finished moving commands around
